+++ docs/src/examples/03-parsimonious-flux-balance.jl
@@ -1,7 +1,7 @@
 
 # # Parsimonious flux balance analysis
 
-# We will use [`parsimonious_flux_balance`](@ref) and
+# We will use [`parsimonious_flux_balance_analysis`](@ref) and
 # [`minimize_metabolic_adjustment`](@ref) to find the optimal flux
 # distribution in the *E. coli* "core" model.
 #
@@ -24,11 +24,11 @@ model = load_model("e_coli_core.json") # load the model
 
 # Use the convenience function to run standard pFBA on
 
-vt = parsimonious_flux_balance(model, Clarabel.Optimizer; modifications = [silence])
+vt = parsimonious_flux_balance_analysis(model, Clarabel.Optimizer; modifications = [silence])
 
 # Or use the piping functionality
 
-model |> parsimonious_flux_balance(Clarabel.Optimizer; modifications = [silence])
+model |> parsimonious_flux_balance_analysis(Clarabel.Optimizer; modifications = [silence])
 
 @test isapprox(vt.objective, 0.87392; atol = TEST_TOLERANCE) #src
 @test sum(x^2 for x in values(vt.fluxes)) < 15000 #src
