+++ docs/src/examples/04-minimization-of-metabolic-adjustment.jl
@@ -1,5 +1,5 @@
 
-# # Minimization of metabolic adjustment
+# # Minimization of metabolic adjustment analysis
 
 # TODO MOMA citation
 
@@ -18,6 +18,6 @@ import Clarabel
 # guessing.
 model = convert(CM.Model, load_model("e_coli_core.json"))
 
-reference_fluxes = parsimonious_flux_balance(model, Clarabel.Optimizer).fluxes
+reference_fluxes = parsimonious_flux_balance_analysis(model, Clarabel.Optimizer).fluxes
 
 # TODO MOMA from here
