+++ src/COBREXA.jl
@@ -31,6 +31,7 @@ include("types.jl")
 include("io.jl")
 include("solver.jl")
 
+# these functions build or extend constrainttrees of metabolic models
 include("builders/core.jl")
 include("builders/genes.jl")
 include("builders/objectives.jl")
@@ -39,11 +40,15 @@ include("builders/thermodynamic.jl")
 include("builders/loopless.jl")
 include("builders/communities.jl")
 
-include("analysis/modifications.jl")
-include("analysis/flux_balance.jl")
-include("analysis/parsimonious_flux_balance.jl")
-include("analysis/minimal_metabolic_adjustment.jl")
+# these are the one shot analysis functions
+include("frontend/flux_balance_analysis.jl")
+include("frontend/parsimonious_flux_balance.jl")
+include("frontend/minimization_of_metabolic_adjustment_analysis.jl")
+include("frontend/enzyme_constrained_flux_balance_analysis.jl")
+include("frontend/loopless_flux_balance_analysis.jl")
+include("frontend/max_min_driving_force_analysis.jl")
 
+include("misc/modifications.jl")
 include("misc/bounds.jl")
 include("misc/utils.jl")
