+++ src/builders/core.jl
@@ -1,6 +1,4 @@
 
-import SparseArrays: sparse
-
 """
 $(TYPEDSIGNATURES)
